tranferMoneyDONE

+++ backend/useraccount/models.py
@@ -95,7 +95,7 @@ class Account(models.Model):
     #id = models.UUIDField(default=uuid.uuid4, editable=False, primary_key=False)
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     account_no = models.CharField(default=accNo() ,max_length=9, editable=False, unique=True)
-    balance = models.PositiveIntegerField(default=0)
+    balance = models.FloatField(default=0)
     qr_code_image = models.ImageField(upload_to='profile/account_qr/', blank=True, null=True)
 
     class Meta:
